Support stats files with stats from different Uncore agents
Previously, the number of uncore agents that we looked at was
hard-coded.  However, for each type of uncore agent, there are
a different number of agents supported.  Automatically detect
the proper number of agents to poll and report on, and if stats
from different agents are showed on the same screen, roll
everything up into a single column.